(nr.parsing.core) bump version to 1.1.0

+++ nr.parsing.core/setup.py
@@ -10,6 +10,9 @@ import sys
 def _tempcopy(src, dst):
   import atexit, shutil
   if not os.path.isfile(dst):
+    if not os.path.isfile(src):
+      print('warning: source file "{}" for destination "{}" does not exist'.format(src, dst))
+      return
     shutil.copyfile(src, dst)
     atexit.register(lambda: os.remove(dst))
 
@@ -31,7 +34,7 @@ test_requirements = [
 
 setuptools.setup(
   name = 'nr.parsing.core',
-  version = '1.0.1',
+  version = '1.1.0',
   author = 'Niklas Rosenstein',
   author_email = 'rosensteinniklas@gmail.com',
   description = 'Simple library to scan or tokenize text.',
